setState関数をそのまま渡さず、App.jsで定義したイベントハンドラ関数を子に渡すように修正
