+++ src/MemoList.js
@@ -2,37 +2,11 @@ import MemoTitle from "./MemoTitle";
 
 export default function MemoList({
   memos,
-  setMemos,
   selectedMemo,
-  setSelectedMemo,
-  setIsEditable,
   isAddingNewMemo,
-  setIsAddingNewMemo,
+  onMemoTitleClick,
+  onAddButtonClick,
 }) {
-  const handleMemoTitleClick = (memo) => {
-    if (selectedMemo !== null && selectedMemo.id === memo.id) {
-      return;
-    }
-
-    setIsEditable(true);
-    setSelectedMemo(memo);
-    setMemos(JSON.parse(localStorage.getItem("memos")));
-    isAddingNewMemo && setIsAddingNewMemo(false);
-  };
-
-  const handleAddButtonClick = () => {
-    const newMemoId = memos.length > 0 ? memos[memos.length - 1].id + 1 : 1;
-    const temporaryNewMemo = {
-      id: newMemoId,
-      content: "新規メモ",
-    };
-    const temporaryNewMemoList = [...memos, temporaryNewMemo];
-    setIsAddingNewMemo(true);
-    setIsEditable(true);
-    setMemos(temporaryNewMemoList);
-    setSelectedMemo(temporaryNewMemo);
-  };
-
   return (
     <div>
       <div>
@@ -46,15 +20,12 @@ export default function MemoList({
                   ? "selected-memo-title"
                   : "memo-title"
               }
-              onClick={() => handleMemoTitleClick(memo)}
+              onClick={() => onMemoTitleClick(memo)}
             />
           ))}
       </div>
       <div>
-        <button
-          onClick={handleAddButtonClick}
-          disabled={isAddingNewMemo && true}
-        >
+        <button onClick={onAddButtonClick} disabled={isAddingNewMemo && true}>
           +
         </button>
       </div>
